app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-multipart==0.0.6
openai==1.3.5
anthropic==0.7.7
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
pydantic[email]==2.5.0
httpx==0.25.2
aiofiles==23.2.1
pandas==2.1.3
numpy==1.25.2
Pillow==10.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic-settings==2.1.0
JWT=1.4.0